requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pdfplumber==0.11.0
PyMuPDF==1.24.5
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.3.0
opencv-python-headless==4.10.0.82
spacy==3.7.5
scikit-learn==1.5.0
numpy==1.26.4
anthropic==0.28.0
transformers==4.41.2
torch==2.3.0
vaderSentiment==3.3.2
textblob==0.18.0
yake==0.4.8
langdetect==1.0.9
pydantic==2.7.1
